refactor: Remove dead deps, modernize imports, deduplicate MyFaiss (Phase 6)
- Remove unused newspaper3k and flask-basicauth dependencies along
  with their 8 transitive packages
- Update deprecated langchain.embeddings.base import to langchain_core
- Replace pytz with stdlib zoneinfo in localization, task_scheduler,
  and state_snapshot (pytz kept in deps for transitive compatibility)
- Extract duplicate MyFaiss class into shared python/helpers/faiss_wrapper.py
- Replace deprecated Flask JSON_SORT_KEYS config with app.json.sort_keys
  in both run_ui.py and run_tunnel.py
