+++ pyproject.toml
@@ -21,7 +21,6 @@ dependencies = [
     "fastmcp==3.0.0b2",
     "flaredantic==0.1.5",
     "flask[async]==3.0.3",
-    "flask-basicauth==0.2.0",
     "GitPython==3.1.43",
     "html2text>=2024.2.26",
     "imapclient>=3.0.1",
@@ -36,7 +35,6 @@ dependencies = [
     "markdownify==1.1.0",
     "mcp>=1.24.0,<2.0",
     "nest-asyncio==1.6.0",
-    "newspaper3k==0.2.8",
     "openai==1.99.5",
     "openai-whisper==20250625",
     "paramiko==3.5.0",
