+++ requirements.txt
@@ -77,9 +77,7 @@ beartype==0.22.9
 beautifulsoup4==4.14.3
     # via
     #   apollos-ai
-    #   feedfinder2
     #   markdownify
-    #   newspaper3k
     #   unstructured
 bidict==0.23.1
     # via python-socketio
@@ -169,8 +167,6 @@ cryptography==46.0.5
     #   requests-ntlm
     #   secretstorage
     #   unstructured-client
-cssselect==1.4.0
-    # via newspaper3k
 csvw==3.7.0
     # via segments
 cuda-bindings==12.9.4 ; platform_machine == 'x86_64' and sys_platform == 'linux'
@@ -250,14 +246,9 @@ fastmcp==3.0.0b2
     # via apollos-ai
 fastuuid==0.14.0
     # via litellm
-feedfinder2==0.0.4
-    # via newspaper3k
-feedparser==6.0.12
-    # via newspaper3k
 filelock==3.20.3
     # via
     #   huggingface-hub
-    #   tldextract
     #   torch
     #   transformers
 filetype==1.2.0
@@ -267,11 +258,8 @@ flaredantic==0.1.5
 flask==3.0.3
     # via
     #   apollos-ai
-    #   flask-basicauth
     #   flask-cors
     #   flask-limiter
-flask-basicauth==0.2.0
-    # via apollos-ai
 flask-cors==6.0.2
     # via apollos-ai
 flask-limiter==4.1.1
@@ -385,7 +373,6 @@ idna==3.11
     #   email-validator
     #   httpx
     #   requests
-    #   tldextract
     #   yarl
 imapclient==3.1.0
     # via apollos-ai
@@ -411,8 +398,6 @@ jeepney==0.9.0 ; sys_platform == 'linux'
     # via
     #   keyring
     #   secretstorage
-jieba3k==0.35.1
-    # via newspaper3k
 jinja2==3.1.6
     # via
     #   flask
@@ -489,7 +474,6 @@ lxml==6.0.2
     # via
     #   exchangelib
     #   lxml-html-clean
-    #   newspaper3k
     #   pikepdf
     #   python-docx
     #   python-pptx
@@ -560,12 +544,8 @@ networkx==3.6.1
     # via
     #   torch
     #   unstructured
-newspaper3k==0.2.8
-    # via apollos-ai
 nltk==3.9.2
-    # via
-    #   newspaper3k
-    #   unstructured
+    # via unstructured
 num2words==0.5.14
     # via misaki
 numba==0.63.1
@@ -728,7 +708,6 @@ pikepdf==10.3.0
 pillow==12.1.0
     # via
     #   matplotlib
-    #   newspaper3k
     #   pdf2image
     #   pi-heif
     #   pikepdf
@@ -1593,7 +1572,6 @@ python-dateutil==2.9.0.post0
     #   botocore
     #   csvw
     #   matplotlib
-    #   newspaper3k
     #   pandas
     #   posthog
     #   unstructured-client
@@ -1642,7 +1620,6 @@ pyyaml==6.0.3
     #   langchain
     #   langchain-community
     #   langchain-core
-    #   newspaper3k
     #   omegaconf
     #   timm
     #   transformers
@@ -1671,7 +1648,6 @@ requests==2.32.5
     #   csvw
     #   docker
     #   exchangelib
-    #   feedfinder2
     #   flaredantic
     #   google-api-core
     #   google-genai
@@ -1681,20 +1657,15 @@ requests==2.32.5
     #   langchain-community
     #   langsmith
     #   msal
-    #   newspaper3k
     #   posthog
-    #   requests-file
     #   requests-ntlm
     #   requests-oauthlib
     #   requests-toolbelt
     #   spacy
     #   tiktoken
-    #   tldextract
     #   transformers
     #   unstructured
     #   weasel
-requests-file==3.0.1
-    # via tldextract
 requests-ntlm==1.3.0
     # via exchangelib
 requests-oauthlib==2.0.0
@@ -1747,8 +1718,6 @@ setuptools==82.0.0
     #   spacy
     #   thinc
     #   torch
-sgmllib3k==1.0.0
-    # via feedparser
 simple-websocket==1.1.0
     # via python-engineio
 simpleeval==1.0.3
@@ -1757,7 +1726,6 @@ simpleeval==1.0.3
     #   casbin
 six==1.17.0
     # via
-    #   feedfinder2
     #   html5lib
     #   langdetect
     #   markdownify
@@ -1830,10 +1798,6 @@ timm==1.0.24
     # via
     #   effdet
     #   unstructured-inference
-tinysegmenter==0.3
-    # via newspaper3k
-tldextract==5.3.1
-    # via newspaper3k
 tokenizers==0.22.2
     # via
     #   litellm
